Initial implementation of last known location on Android
